Feature/embedded headless pbft

+++ nekoyume/Assets/Planetarium/Nekoyume/Editor/HeadlessTool.cs
@@ -1,8 +1,10 @@
 // See also: HeadlessHelper.cs
+
 using System;
 using System.Diagnostics;
 using System.IO;
 using System.Runtime.InteropServices;
+using Libplanet.Crypto;
 using Nekoyume.BlockChain;
 using UnityEditor;
 using UnityEngine;
@@ -62,6 +64,34 @@ namespace Planetarium.Nekoyume.Editor
             var startInfo = new ProcessStartInfo
             {
                 FileName = "git",
+                Arguments = "reset HEAD --hard",
+                WorkingDirectory = _headlessPath,
+            };
+            process = Process.Start(startInfo);
+            process.WaitForExit();
+
+            startInfo = new ProcessStartInfo
+            {
+                FileName = "git",
+                Arguments = "checkout development",
+                WorkingDirectory = _headlessPath,
+            };
+            process = Process.Start(startInfo);
+            process.WaitForExit();
+
+            startInfo = new ProcessStartInfo
+            {
+                FileName = "git",
+                Arguments = "pull origin development",
+                WorkingDirectory = _headlessPath,
+            };
+            process = Process.Start(startInfo);
+            process.WaitForExit();
+            Debug.Log("Pull latest commits from origin.");
+
+            startInfo = new ProcessStartInfo
+            {
+                FileName = "git",
                 Arguments = @"submodule update --init --recursive NineChronicles.RPC.Shared",
                 WorkingDirectory = _headlessPath,
             };
@@ -112,7 +142,16 @@ namespace Planetarium.Nekoyume.Editor
         {
             // TODO: Select option to use local genesis or download genesis from URL (like use mainnet genesis)
             Debug.Log($"Delete and create genesis-block to {BlockManager.GenesisBlockPath()}");
-            Directory.Delete(Path.Combine(_docsRoot, "planetarium", _storeName), true);
+            if (Directory.Exists(Path.Combine(_docsRoot, "planetarium", _storeName)))
+            {
+                Directory.Delete(Path.Combine(_docsRoot, "planetarium", _storeName), true);
+            }
+
+            if (!Directory.Exists(Path.Combine(_docsRoot, "planetarium")))
+            {
+                Directory.CreateDirectory(Path.Combine(_docsRoot, "planetarium"));
+            }
+
             LibplanetEditor.DeleteAllEditorAndMakeGenesisBlock();
         }
 
@@ -155,6 +194,14 @@ namespace Planetarium.Nekoyume.Editor
                 Arguments =
                     $"run -c DevEx --project NineChronicles.Headless.Executable -C appsettings.local.json --genesis-block-path {Path.Combine(_genesisPath, "genesis-block")} --store-path {Path.Combine(_docsRoot, "planetarium", _storeName)} --store-type memory",
             };
+
+            if (PlayerPrefs.HasKey("initialValidator"))
+            {
+                var pkHex = PlayerPrefs.GetString("initialValidator");
+                startInfo.Arguments +=
+                    $" --miner-private-key {pkHex} --consensus-private-key {pkHex} --consensus-seed {new PrivateKey(pkHex).PublicKey},localhost,60000";
+            }
+
             Debug.Log(startInfo.Arguments);
             startInfo.WorkingDirectory = _headlessPath;
             Debug.Log($"WorkingDirectory: {startInfo.WorkingDirectory}");
